aggiornamento automatico performance cittadino e stato cassonetto

+++ CitizenAccountManagement/src/main/java/it/unisalento/pas/citizenaccountmanagement/domain/Citizen.java
@@ -13,7 +13,7 @@ public class Citizen {
     String email;
     String comune;
     int da_sensibilizzare; /* 1 True, 0 False*/
-    int performance; /* in percentuale */
+    float performance; /* in percentuale */
     String password;
 
     public String getPassword() {
@@ -72,11 +72,11 @@ public class Citizen {
         this.da_sensibilizzare = da_sensibilizzare;
     }
 
-    public int getPerformance() {
+    public float getPerformance() {
         return performance;
     }
 
-    public void setPerformance(int performance) {
+    public void setPerformance(float performance) {
         this.performance = performance;
     }
 }
